Refactor AppComponent creating subcomponent PlaylistCreationTool
